app/services/s3_service.py
import boto3
import os
from dotenv import load_dotenv

load_dotenv()

AWS_ACCESS_KEY_ID = os.getenv("AWS_ACCESS_KEY_ID")
AWS_SECRET_ACCESS_KEY = os.getenv("AWS_SECRET_ACCESS_KEY")
AWS_REGION = os.getenv("AWS_REGION")
AWS_S3_BUCKET = os.getenv("AWS_S3_BUCKET")

s3_client = boto3.client(
    "s3",
    region_name=AWS_REGION,
    aws_access_key_id=AWS_ACCESS_KEY_ID,
    aws_secret_access_key=AWS_SECRET_ACCESS_KEY,
)

def upload_file_to_s3(file_path: str, s3_key: str):
    try:
        s3_client.upload_file(file_path, AWS_S3_BUCKET, s3_key)
        return f"s3://{AWS_S3_BUCKET}/{s3_key}"
    except Exception as e:
        raise RuntimeError(f"Failed to upload to S3: {str(e)}")

