requirements.txt
fastapi
uvicorn
python-multipart
transformers
boto3
python-dotenv
pdf2image
pytesseract
Pillow
